@@ -5,19 +5,20 @@
   "dependencies": {
     "connected-react-router": "^4.4.1",
     "d3-scale": "^2.1.0",
-    "konva": "^2.3.0",
+    "konva": "^2.5.1",
     "moment": "^2.22.2",
-    "office-ui-fabric-core": "^9.6.0",
-    "office-ui-fabric-react": "^6.50.1",
+    "office-ui-fabric-core": "^9.6.1",
+    "office-ui-fabric-react": "^6.106.3",
     "prop-types": "^15.6.2",
-    "react": "^16.4.1",
+    "react": "^16.6.3",
+    "react-app-polyfill": "^0.1.3",
     "react-big-calendar": "^0.19.2",
-    "react-dom": "^16.4.2",
-    "react-konva": "^1.7.14",
+    "react-dom": "^16.6.3",
+    "react-konva": "^16.6.31",
     "react-redux": "^5.0.7",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "^1.1.4",
+    "react-scripts": "^2.1.1",
     "react-transition-group": "^2.4.0",
     "redux": "^4.0.0",
     "redux-logger": "^3.0.6",
@@ -31,7 +32,7 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "devDependencies": {
@@ -43,5 +44,11 @@
     "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-react": "^7.11.1",
     "jest": "^23.5.0"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
